Modify Pokérus annoucements from scheduler -> http
Previously a scheduler was set up to announce the Pokérus host every X
minutes. This was very buggy and often the scheduler would run twice.

Now Pokérus is announced when Luna receives a request on a specific
server. The server runs on localhost and the port is passed in the
command line options e.g. -pokerusserver ":8080"